requirements.txt
Django==4.2.7
Pillow==10.4.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
django-allauth==0.57.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
reportlab==4.0.7
